feat(gedcom): Add application wrapper for GEDCOM data
This is another blob of code I worked with for a couple of months in
another private personal project.

Add some types to represent GEDCOM records. The imported library is
mostly written for GEDCOM v7, but in practice I've found it works
fine for GEDCOM v5.5.1 as well for this application's needs.

Add sample GEDCOM files for testing. These were copied from
https://github.com/findmypast/gedcom-samples.