cmd: Reorganize relate subcmd
Makes the use of fzf optional; not everyone has this installed. If it's
not installed, then suggest it. Offer a simple alternative of directly
inputting the IDs. Update usage func w/ examples. Do some refactoring of
the fzf functionality so that any kind of specific data formatting is
more isolated.