test(gedcom): Add, update tests
Change package for PersonalName tests to simplify construction of
different scenarios and checking source citations. It's easier to rely
on the GEDCOM specification than it is to rely on minutiae of method
signatures and types from the the golang code.

Extract some common assertions, like comparisons for []string and for
map[string]string, into helper funcs. Reasoning for not choosing
generics for those helper funcs is in a code comment.

Organize tests