Improve draw functionality
Move some logic out of the templates. Much easier construct data in
regular golang than in the templating language.

Use the elk rendering engine, which is only available for mermaid
versions >= v9.4. As of 2023-10, the docs say this is experimental and
that it must be enabled. For family trees, the elk rendering engine is
better than the default b/c lines are easier to follow.

Added some CLI options:
- Specify flowchart direction
- Omit display ID by default

Updated the union nodes so they're easier to read and distinguish from
the person nodes.
Add person initials for union nodes labels. Less space to take up.
Having a brief reference to the partners of a union inside the node
label is really helpful when the family tree starts getting dense.

Update tests