Change plurality of dates in GEDCOM data
It's not uncommon for individuals in GEDCOM to have multiple views of
the birthdate due to sources having entirely different dates or
different granularities of dates.
Also, when there is a collection of the same tags, the first one is
considered the "preferred" value. So pick that one most of the time.