feat: Add explore-data functionality
Add a shell script, explore-data.sh, with 2 subcommands, show and
relate. The relate subcommand already exists, but show is new. The show
subcommand is meant to imitate a "family group sheet" view, as it's
sometimes known. This is meant to give you a compact view of somebody's
lifespan, and those of their immediate families.

Also start using fzf directly, rather than have the golang code start
the process. I've found that it's much easier to modify a script that
invokes fzf than it is to modify the golang code and have it invoke fzf.

There's definitely a trade off here: the input data is read into fzf
twice now. Once to generate the lines data (fzf uses this input so it
can help you filter down to a person), and another time to perform the
subcommand, which is either "show" details on a person, or "relate"
these two people. I'm doing this b/c, at this time, it's easier to build
an interactive UI this way. While there are now some redundant reads, I
don't think it'll be that noticeable. After all, this program is meant
for just one user in a terminal. One option for mitigation is to use
bkt, https://github.com/dimo414/bkt, for caching lookups.

In package cmd, adjust data shaping so that it can be fed into fzf for
quick interactive lookups, and output to the user via the terminal. The
user-facing output is either JSON or is shaped by the library,
github.com/charmbracelet/lipgloss, so it's more human-readable.

To feed into fzf, format the GED data into a "lines" format. The
"relate" subcommand, already does this, but now it's exposed as its own
subcommand, "parse to-lines".

Other misc things:

- Update justfile to not print command for run recipe
- Add justfile recipes for maintaining shell scripts
- Add graceful fallback behavior when user does not have fzf
- feat: Add gedcom.PersonalName.String method
- refactor: Adjust entity.Date.String method