fix typo, fix log init bug
The log init bug allowed for the log.Init function to be called multiple
times with different results. This was an improper use of sync.OnceFunc.
The intended use requires access to an input slog.Handler; it doesn't
make sense to redefine the initialization upon every call to log.Init.